batchservice unit tests
