--- src/test/groovy/au/org/ala/images/BatchFileUploadSpec.groovy
@@ -1,16 +1,0 @@
-package au.org.ala.images
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class BatchFileUploadSpec extends Specification implements DomainUnitTest<BatchFileUpload> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-    }
-}
